fix: ゲームセット時にエラーになるのを修正した。

+++ cogs/minesweeper.py
@@ -73,6 +73,7 @@ class Mines(commands.Cog):
                 kek = game.set(int(x), int(y))
             except ValueError:
                 pass
+            if kek in (410, 301): break
             try:
                 await mes.delete()
             except BaseException:
